# evidencias/conformidade_2026-10-02/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API de Agendamento de Consultas",
    "description": "Assessment: use somente dados ficticios neste ambiente local.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/consultas": {
      "post": {
        "tags": [
          "Consultas"
        ],
        "summary": "Criar Consulta",
        "operationId": "criar_consulta_consultas_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "JWTBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultaWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "413": {
            "description": "Corpo maior que o limite configurado."
          },
          "401": {
            "description": "Credencial ausente, invalida ou revogada."
          },
          "403": {
            "description": "Papel ou vinculo nao autorizado."
          },
          "503": {
            "description": "Persistencia ou auditoria indisponivel."
          },
          "429": {
            "description": "Orcamento esgotado; consultar Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "409": {
            "description": "Conflito de persistencia ou estado."
          }
        }
      },
      "get": {
        "tags": [
          "Consultas"
        ],
        "summary": "Listar Consultas",
        "operationId": "listar_consultas_consultas_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          },
          {
            "JWTBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2147483647,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsultaRead"
                  },
                  "title": "Response Listar Consultas Consultas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "413": {
            "description": "Corpo maior que o limite configurado."
          },
          "401": {
            "description": "Credencial ausente, invalida ou revogada."
          },
          "403": {
            "description": "Papel ou vinculo nao autorizado."
          },
          "503": {
            "description": "Persistencia ou auditoria indisponivel."
          },
          "429": {
            "description": "Orcamento esgotado; consultar Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          }
        }
      }
    },
...
    "/integracoes/token": {
      "post": {
        "tags": [
          "Laboratorio M2M"
        ],
        "summary": "Machine Token",
        "operationId": "machine_token_integracoes_token_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineTokenRead"
                }
              }
            }
          },
          "413": {
            "description": "Corpo maior que o limite configurado."
          },
          "401": {
            "description": "Credencial ausente, invalida ou revogada."
          },
          "403": {
            "description": "Papel ou vinculo nao autorizado."
          },
          "503": {
            "description": "Persistencia ou auditoria indisponivel."
          },
          "429": {
            "description": "Orcamento esgotado; consultar Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "400": {
            "description": "invalid_request ou invalid_scope."
          }
        },
        "security": [
          {
            "ClientBasic": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "grant_type": {
                    "const": "client_credentials",
                    "title": "Grant Type",
                    "type": "string"
                  },
                  "scope": {
                    "default": "disponibilidade:read",
                    "maxLength": 200,
                    "title": "Scope",
                    "type": "string"
                  }
                },
                "required": [
                  "grant_type"
                ],
                "title": "ClientCredentialsForm",
                "type": "object"
              }
            }
          }
        },
        "description": "Client credentials com HTTP Basic. Nunca usar credenciais humanas. Scope omitido concede disponibilidade:read; scope vazio/desconhecido e rejeitado. Token expira em 300 segundos, sem refresh token; requer TLS fora de loopback."
      }
    },
    "/integracoes/disponibilidade": {
      "get": {
        "tags": [
          "Laboratorio M2M"
        ],
        "summary": "Availability",
        "operationId": "availability_integracoes_disponibilidade_get",
        "parameters": [
          {
            "name": "dia",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Dia"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HorarioRead"
                  },
                  "title": "Response Availability Integracoes Disponibilidade Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "413": {
            "description": "Corpo maior que o limite configurado."
          },
          "401": {
            "description": "Credencial ausente, invalida ou revogada."
          },
          "403": {
            "description": "Papel ou vinculo nao autorizado."
          },
          "503": {
            "description": "Persistencia ou auditoria indisponivel."
          },
          "429": {
            "description": "Orcamento esgotado; consultar Retry-After.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          }
        },
        "security": [
          {
            "LaboratorioOAuth2": [
              "disponibilidade:read"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ConsultaRead": {
        "properties": {
          "id": {
            "type": "integer",
            "maximum": 2147483647.0,
            "exclusiveMinimum": 0.0,
            "title": "Id"
          },
          "paciente_id": {
            "type": "integer",
            "maximum": 2147483647.0,
            "exclusiveMinimum": 0.0,
            "title": "Paciente Id"
          },
          "profissional_id": {
            "type": "integer",
            "maximum": 2147483647.0,
            "exclusiveMinimum": 0.0,
            "title": "Profissional Id"
          },
          "data_hora": {
            "type": "string",
            "format": "date-time",
            "title": "Data Hora"
          },
          "status": {
            "$ref": "#/components/schemas/StatusConsulta"
          },
          "observacao": {
            "type": "string",
            "title": "Observacao"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "paciente_id",
          "profissional_id",
          "data_hora",
          "status",
          "observacao"
        ],
        "title": "ConsultaRead"
      },
      "ConsultaWrite": {
        "properties": {
          "paciente_id": {
            "type": "integer",
            "maximum": 2147483647.0,
            "exclusiveMinimum": 0.0,
            "title": "Paciente Id"
          },
          "profissional_id": {
            "type": "integer",
            "maximum": 2147483647.0,
            "exclusiveMinimum": 0.0,
            "title": "Profissional Id"
          },
          "data_hora": {
            "type": "string",
            "format": "date-time",
            "title": "Data Hora"
          },
          "status": {
            "$ref": "#/components/schemas/StatusConsulta",
            "default": "agendada"
          },
          "observacao": {
            "type": "string",
            "maxLength": 500,
            "title": "Observacao",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "paciente_id",
          "profissional_id",
          "data_hora"
        ],
        "title": "ConsultaWrite",
        "description": "Contrato para criacao e substituicao completa (PUT)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HorarioRead": {
        "properties": {
          "profissional_id": {
            "type": "integer",
            "maximum": 2147483647.0,
            "exclusiveMinimum": 0.0,
            "title": "Profissional Id"
          },
          "inicio": {
            "type": "string",
            "format": "date-time",
            "title": "Inicio"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "profissional_id",
          "inicio"
        ],
        "title": "HorarioRead"
      },
...
      "MachineTokenRead": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "const": "bearer",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "default": 300
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "access_token",
          "scope"
        ],
        "title": "MachineTokenRead"
      },
      "StatusConsulta": {
        "type": "string",
        "enum": [
          "agendada",
          "cancelada",
          "realizada"
        ],
        "title": "StatusConsulta"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/token"
          }
        }
      },
      "ClientBasic": {
        "type": "http",
        "scheme": "basic"
      },
      "LaboratorioOAuth2": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "integracoes/token",
            "scopes": {
              "disponibilidade:read": "Ler somente horarios explicitamente ofertados."
            }
          }
        }
      },
      "JWTBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtido no login ou apos confirmar MFA."
      },
      "AgendaCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agenda_session",
        "description": "Cookie aceito somente no GET /agenda."
      }
    }
  }
}